web_interface_service/requirements.txt
gradio==4.29.0
psycopg2-binary==2.9.6
pillow==10.3.0
numpy==1.26.4
pandas==2.2.2